// deliveryClient/models/vehicle.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0'
 */
export type Vehicle = IContentItem<{
  name: Elements.TextElement;
  photo: Elements.AssetsElement;
  workshops: Elements.LinkedItemsElement<IContentItem>;
  price_for_90_minutes: Elements.NumberElement;
  type: Elements.MultipleChoiceElement;
}>;